+++ src/redturtle/importer/base/transmogrifier/deserializers/datetime.py
@@ -2,6 +2,7 @@
 from DateTime import DateTime
 from datetime import datetime
 from zope.schema.interfaces import IDatetime
+from zope.schema.interfaces import IDate
 from plone.app.event.base import default_timezone
 from redturtle.importer.base.interfaces import IDeserializer
 from zope.interface import implementer
@@ -24,8 +25,6 @@ class DatetimeDeserializer(object):
     ):
         if value == "None":
             return None
-        if isinstance(value, datetime):
-            value = value.date()
         if isinstance(value, six.string_types):
             # Fix some rare use case
             if "Universal" in value:
@@ -39,6 +38,35 @@ class DatetimeDeserializer(object):
                 value = value.astimezone(tz_default)
             except ValueError:
                 value = DateTime(value)
+        try:
+            self.field.validate(value)
+        except Exception as e:
+            if not disable_constraints:
+                raise e
+            else:
+                if logger:
+                    logger(
+                        "%s is invalid in %s: %s"
+                        % (self.field.__name__, item["_path"], e)
+                    )
+
+        return value
+
+
+@implementer(IDeserializer)
+@adapter(IDate, Interface)
+class DateDeserializer(object):
+    def __init__(self, field, context):
+        self.field = field
+        self.context = context
+
+    def __call__(
+        self, value, filestore, item, disable_constraints=False, logger=None
+    ):
+        if value == "None":
+            return None
+        if isinstance(value, six.string_types):
+            value = DateTime(value).asdatetime().date()
         try:
             self.field.validate(value)
         except Exception as e: